enum MovieGrade 중복제거
